re-org the dir
