+++ Cargo.toml
@@ -18,7 +18,7 @@
 [workspace]
 exclude = ["datafusion-cli"]
 members = ["datafusion/common", "datafusion/core", "datafusion/expr", "datafusion/execution", "datafusion/optimizer", "datafusion/physical-expr", "datafusion/proto", "datafusion/proto/gen", "datafusion/sql", "datafusion/sqllogictest", "datafusion/substrait", "datafusion-examples", "test-utils", "benchmarks", 
-"extension/functions/presto",
+"extension/functions",
 ]
 resolver = "2"
