+++ extension/functions/Cargo.toml
@@ -16,7 +16,7 @@
 # under the License.
 
 [package]
-name = "datafusion-extension-functions-presto"
+name = "datafusion-extension-functions"
 authors.workspace = true
 edition.workspace = true
 homepage.workspace = true
@@ -29,10 +29,16 @@ version.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-datafusion = { path = "../../../datafusion/core" }
-datafusion-common = { path = "../../../datafusion/common" }
-datafusion-expr = { path = "../../../datafusion/expr" }
+datafusion = { path = "../../datafusion/core" }
+datafusion-common = { path = "../../datafusion/common" }
+datafusion-expr = { path = "../../datafusion/expr" }
 arrow = { workspace = true }
 #arrow-flight = { workspace = true }
 #arrow-schema = { workspace = true }
 tokio = "1.33.0"
+
+[features]
+example = []
+ansi = []
+presto = ["ansi"]
+postgres = ["ansi"]
