+++ extension/functions/src/example.rs
@@ -88,9 +88,9 @@ impl ScalarFunctionDef for MultiplyTwoFunction {
 }
 
 // Function package declaration
-pub struct TestFunctionPackage;
+pub struct FunctionPackage;
 
-impl ScalarFunctionPackage for TestFunctionPackage {
+impl ScalarFunctionPackage for FunctionPackage {
     fn functions(&self) -> Vec<Box<dyn ScalarFunctionDef>> {
         vec![Box::new(AddOneFunction), Box::new(MultiplyTwoFunction)]
     }
@@ -98,67 +98,14 @@ impl ScalarFunctionPackage for TestFunctionPackage {
 
 #[cfg(test)]
 mod test {
-    use arrow::{
-        array::ArrayRef, record_batch::RecordBatch, util::display::array_value_to_string,
-    };
     use datafusion::error::Result;
     use datafusion::prelude::SessionContext;
     use tokio;
 
-    use crate::TestFunctionPackage;
 
-    /// Execute query and return result set as 2-d table of Vecs
-    /// `result[row][column]`
-    async fn execute(ctx: &SessionContext, sql: &str) -> Vec<Vec<String>> {
-        result_vec(&execute_to_batches(ctx, sql).await)
-    }
+    use crate::utils::{test_expression, execute};
 
-    /// Specialised String representation
-    fn col_str(column: &ArrayRef, row_index: usize) -> String {
-        if column.is_null(row_index) {
-            return "NULL".to_string();
-        }
-
-        array_value_to_string(column, row_index)
-            .ok()
-            .unwrap_or_else(|| "???".to_string())
-    }
-    /// Converts the results into a 2d array of strings, `result[row][column]`
-    /// Special cases nulls to NULL for testing
-    fn result_vec(results: &[RecordBatch]) -> Vec<Vec<String>> {
-        let mut result = vec![];
-        for batch in results {
-            for row_index in 0..batch.num_rows() {
-                let row_vec = batch
-                    .columns()
-                    .iter()
-                    .map(|column| col_str(column, row_index))
-                    .collect();
-                result.push(row_vec);
-            }
-        }
-        result
-    }
-
-    /// Execute query and return results as a Vec of RecordBatches
-    async fn execute_to_batches(ctx: &SessionContext, sql: &str) -> Vec<RecordBatch> {
-        let df = ctx.sql(sql).await.unwrap();
-
-        // optimize just for check schema don't change during optimization.
-        df.clone().into_optimized_plan().unwrap();
-
-        df.collect().await.unwrap()
-    }
-
-    macro_rules! test_expression {
-        ($SQL:expr, $EXPECTED:expr) => {
-            let ctx = SessionContext::new();
-            ctx.register_scalar_function_package(Box::new(TestFunctionPackage));
-            let sql = format!("SELECT {}", $SQL);
-            let actual = execute(&ctx, sql.as_str()).await;
-            assert_eq!(actual[0][0], $EXPECTED);
-        };
-    }
+    use super::FunctionPackage;
 
     #[tokio::test]
     async fn test_add_one() -> Result<()> {
